- some refactoring in EStore classes - define DelegatedEStore as abstract
